updated some bugs
-planets didn't check if sectors were full
-after reproducing didn't set turns left on new entity (might be causing my rng errors)

+++ simodyssey2-starter/simodyssey2/simodyssey2/model/etf_model.e
@@ -336,7 +336,7 @@ feature --support
 				if not entity.item.is_explorer then
 					if entity.item.turns_left = 0 and not entity.item.is_dead then
 						--special case for planet
-						if entity.item.is_planet then
+						if entity.item.is_planet and entity.item.location.has_star then
 							--planet can't kill anything so don't need to check if
 							--any items were returns by behave
 							create entities_killed.make_from_iterable (entity.item.behave)
